+++ Breakout/DebugEffect.cpp
@@ -1,0 +1,20 @@
+#include "DebugEffect.h"
+
+#include <iostream>
+
+DebugEffect::DebugEffect(const std::string& _debugMessage) :_debugMessage{ _debugMessage }
+{
+}
+
+void DebugEffect::ApplyEffect(float duration)
+{
+	BaseEffect::ApplyEffect(duration);
+	std::cout << _debugMessage << '\n';
+}
+
+
+
+BaseEffect* DebugEffect::CopyEffect()
+{
+	return new DebugEffect(_debugMessage);
+}
