+++ Breakout/FireBallEffect.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "BaseEffect.h"
+
+class Ball;
+
+class FireBallEffect : public BaseEffect
+{
+public:
+	FireBallEffect(Ball* ball);
+
+	void ApplyEffect(float duration) override;
+	void EndEffect() override;
+
+	std::string GetName() override { return "Fire ball"; }
+
+	BaseEffect* CopyEffect();
+private:
+	Ball* _ball;
+};
